+++ source/Paging.hpp
@@ -12,7 +12,7 @@ class Paging
 
     static void init();
 
-    static void mapPageTo(void* vaddr, uint64_t page);
+    static void mapPageTo(void* vaddr, uintptr_t page);
     static void mapPage(void* vaddr, void** paddr = nullptr);
     static void unmapPage(void* vaddr);
